Adding user validation endpoint

+++ src/main/java/com/hardToFind/resources/UserResource.java
@@ -9,6 +9,7 @@ import javax.ws.rs.*;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.MediaType;
 import java.util.List;
+import java.util.Scanner;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
@@ -25,6 +26,18 @@ public class UserResource {
         // this.configuration = configuration;
         this.counter = new AtomicLong();
     }
+
+    @Path("/validate")
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    @Timed
+    public List<User> getValidate(@QueryParam("id") String id, @Context DSLContext database) {
+        UserService userService = new UserService(database);
+        if(id==null || !isInteger(id)) throw new WebApplicationException("No valid id provided");
+        userService.setUserValidated(Integer.parseInt(id));
+        return userService.getUser(Integer.parseInt(id));
+    }
+
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
@@ -56,10 +69,20 @@ public class UserResource {
     @Produces(MediaType.APPLICATION_JSON)
     @Timed
     public Boolean deleteUser(@QueryParam("id") String id, @Context DSLContext database){
-        if(id==null) throw new WebApplicationException("No valid id provided");
+        if(id==null || !isInteger(id)) throw new WebApplicationException("No valid id provided");
         UserService userService = new UserService(database);
         return userService.deleteUser(Integer.parseInt(id));
 
+    }
+
+    private boolean isInteger(String s) {
+        int radix = 10;
+        Scanner sc = new Scanner(s.trim());
+        if(!sc.hasNextInt(radix)) return false;
+        // we know it starts with a valid int, now make sure
+        // there's nothing left!
+        sc.nextInt(radix);
+        return !sc.hasNext();
     }
 
 }
